+++ packages/utils/src/index.ts
@@ -13,6 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+import { ListenerPayload, NotificationListener } from '@optimizely/optimizely-sdk/lib/shared_types';
 import { v4 } from 'uuid';
 
 export function generateUUID(): string {
@@ -154,4 +155,8 @@ export enum NOTIFICATION_TYPES {
 
 export interface NotificationCenter {
   sendNotifications(notificationType: NOTIFICATION_TYPES, notificationData?: any): void
+  addNotificationListener<T extends ListenerPayload>(
+    notificationType: string,
+    callback: NotificationListener<T>
+  ): number;
 }